+++ packages/functions-utils/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [5.2.18](https://github.com/netlify/build/compare/functions-utils-v5.2.17...functions-utils-v5.2.18) (2023-07-10)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @netlify/zip-it-and-ship-it to v9.13.0 ([#5155](https://github.com/netlify/build/issues/5155)) ([4044aa2](https://github.com/netlify/build/commit/4044aa2e862465e44c175f1534307522308f93c6))
+
+## [5.2.17](https://github.com/netlify/build/compare/functions-utils-v5.2.16...functions-utils-v5.2.17) (2023-07-07)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @netlify/zip-it-and-ship-it to v9.12.2 ([#5144](https://github.com/netlify/build/issues/5144)) ([9f8ad3b](https://github.com/netlify/build/commit/9f8ad3b69bbcb457dae0301692fcbcdfbad1c970))
+
+## [5.2.16](https://github.com/netlify/build/compare/functions-utils-v5.2.15...functions-utils-v5.2.16) (2023-07-06)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @netlify/zip-it-and-ship-it to v9.12.1 ([#5130](https://github.com/netlify/build/issues/5130)) ([55293df](https://github.com/netlify/build/commit/55293df75b8875fd006ea59a1f8ad036c7915a6e))
+
+## [5.2.15](https://github.com/netlify/build/compare/functions-utils-v5.2.14...functions-utils-v5.2.15) (2023-07-05)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @netlify/zip-it-and-ship-it to v9.12.0 ([#5107](https://github.com/netlify/build/issues/5107)) ([bff7945](https://github.com/netlify/build/commit/bff7945577b07a071dd53cacd22b0ec0ba152f50))
+
 ## [5.2.14](https://github.com/netlify/build/compare/functions-utils-v5.2.13...functions-utils-v5.2.14) (2023-06-29)